create backgrounds and write twitter handle on em
